feat: Init 'wallet' endpoint

+++ internal/datasources/records/record.product.go
@@ -10,7 +10,7 @@ type Product struct {
 	Id          int        `db:"product_id"`
 	Name        string     `db:"name"`
 	Description string     `db:"description"`
-	Price       float32    `db:"price"`
+	Price       float64    `db:"price"`
 	Stock       int        `db:"stock"`
 	CreatedAt   time.Time  `db:"created_at"`
 	UpdatedAt   *time.Time `db:"updated_at"`
